+++ src/app/loadout/loadout-ui/LoadoutParametersDisplay.tsx
@@ -1,7 +1,6 @@
 import { LoadoutParameters } from '@destinyitemmanager/dim-api-types';
 import BungieImage from 'app/dim-ui/BungieImage';
 import { t } from 'app/i18next-t';
-import { SelectedArmorUpgrade } from 'app/loadout-builder/filter/ArmorUpgradePicker';
 import ExoticArmorChoice from 'app/loadout-builder/filter/ExoticArmorChoice';
 import { useD2Definitions } from 'app/manifest/selectors';
 import { AppIcon, searchIcon } from 'app/shell/icons';
@@ -10,7 +9,7 @@ import styles from './LoadoutParametersDisplay.m.scss';
 
 export default function LoadoutParametersDisplay({ params }: { params: LoadoutParameters }) {
   const defs = useD2Definitions()!;
-  const { query, exoticArmorHash, upgradeSpendTier, statConstraints, lockItemEnergyType } = params;
+  const { query, exoticArmorHash, statConstraints } = params;
   const show =
     params.query ||
     params.exoticArmorHash ||
@@ -31,15 +30,6 @@ export default function LoadoutParametersDisplay({ params }: { params: LoadoutPa
       {exoticArmorHash && (
         <div className={styles.loExotic}>
           <ExoticArmorChoice lockedExoticHash={exoticArmorHash} />
-        </div>
-      )}
-      {upgradeSpendTier !== undefined && (
-        <div className={styles.loSpendTier}>
-          <SelectedArmorUpgrade
-            defs={defs}
-            upgradeSpendTier={upgradeSpendTier}
-            lockItemEnergyType={lockItemEnergyType ?? false}
-          />
         </div>
       )}
       {statConstraints && (